+++ app/System/Language/En/Welcome.php
@@ -63,4 +63,8 @@ return [
 
 	// Friends
 	'uap_friends' => "Friends",
+
+	// Terms and Privacy
+	'terms_title' => "Terms",
+	'privacy_title' => "Privacy",
 ];